Fix warning.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@42903 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk/region.cpp
@@ -459,13 +459,13 @@ wxRect wxRegionIterator::GetRect() const
 wxRegionIterator& wxRegionIterator::operator=(const wxRegionIterator& ri)
 {
     wxDELETEA(m_rects);
-
+    
     m_current = ri.m_current;
     m_numRects = ri.m_numRects;
     if ( m_numRects )
     {
         m_rects = new wxRect[m_numRects];
-        for ( long n = 0; n < m_numRects; n++ )
+        for ( unsigned int n = 0; n < m_numRects; n++ )
             m_rects[n] = ri.m_rects[n];
     }
     else
